feat : 컴포넌트 분리, 추가디자인 수정
