+++ lib/utils/getColor.dart
@@ -1,0 +1,14 @@
+// Setting spreading color when tapping on the button
+import 'package:flutter/material.dart';
+
+MaterialStateProperty<Color> getColor(Color color, Color colorPressed) {
+  getColor(Set<MaterialState> states) {
+    if (states.contains(MaterialState.pressed)) {
+      return colorPressed;
+    } else {
+      return color;
+    }
+  }
+
+  return MaterialStateProperty.resolveWith(getColor);
+}
